# backendnew/docs/swagger.yaml
basePath: /
definitions:
  handler.SetStatusRequest:
    properties:
      asset_id:
        description: Asset ID (empty for global items)
        example: asset-123
        type: string
      item_id:
        description: Checklist item template ID
        example: security-policy-001
        type: string
      notes:
        description: Optional notes
        example: Verified during security audit
        type: string
      status:
        description: 'Status: yes, no, or na'
        enum:
        - "yes"
        - "no"
        - na
        example: "yes"
        type: string
    type: object
  handler.UploadTemplatesRequest:
    properties:
      templates:
        description: Array of checklist templates to upload
        items:
          $ref: '#/definitions/model.ChecklistItemTemplate'
        type: array
    type: object
  model.AssetCoverage:
    properties:
      asset_id:
        type: string
      asset_type:
        type: string
      asset_value:
        type: string
      notes:
        type: string
      status:
        description: '"yes", "no" (excludes "na")'
        type: string
      updated_at:
        type: string
    type: object
  model.ChecklistItemGuide:
    properties:
      acceptance_summary:
        type: string
      faq:
        items:
          $ref: '#/definitions/model.FAQ'
        type: array
      non_technical_steps:
        items:
          type: string
        type: array
      scope_caveats:
        type: string
    type: object
  model.ChecklistItemInfo:
    properties:
      guide:
        $ref: '#/definitions/model.ChecklistItemGuide'
      law_refs:
        items:
          type: string
        type: array
      legal:
        $ref: '#/definitions/model.ChecklistItemLegal'
      priority:
        description: '"must", "should", "may"'
        type: string
      resources:
        items:
          $ref: '#/definitions/model.ChecklistItemResource'
        type: array
      risks:
        $ref: '#/definitions/model.ChecklistItemRisks'
      what_it_means:
        type: string
      why_it_matters:
        type: string
    type: object
  model.ChecklistItemLegal:
    properties:
      article_refs:
        items:
          type: string
        type: array
      quotes:
        items:
          $ref: '#/definitions/model.Quote'
        type: array
      requirement_summary:
        type: string
    type: object
  model.ChecklistItemResource:
    properties:
      title:
        type: string
      url:
        type: string
    type: object
  model.ChecklistItemRisks:
    properties:
      attack_vectors:
        items:
          type: string
        type: array
      potential_impact:
        items:
          type: string
        type: array
    type: object
  model.ChecklistItemTemplate:
    properties:
      asset_types:
        description: Applicable asset types if scope is "asset"
        items:
          type: string
        type: array
      category:
        type: string
      description:
        type: string
      evidence_rules:
        description: Rules for auto-derivation
        items:
          $ref: '#/definitions/model.EvidenceRule'
        type: array
      help_text:
        description: Extended metadata fields for rich UI display
        type: string
      id:
        type: string
      info:
        $ref: '#/definitions/model.ChecklistItemInfo'
      kind:
        description: '"manual" or "auto"'
        type: string
      read_only:
        type: boolean
      recommendation:
        type: string
      required:
        type: boolean
      scope:
        description: '"global", "asset", or "business_unit"'
        type: string
      script_controlled:
        description: Can be controlled by Lua scripts
        type: boolean
      title:
        type: string
      why_matters:
        type: string
    type: object
  model.DerivedChecklistItem:
    properties:
      asset_types:
        description: Applicable asset types if scope is "asset"
        items:
          type: string
        type: array
      attachments:
        description: File attachment IDs
        items:
          type: string
        type: array
      category:
        type: string
      covered_assets:
        description: Assets covered by this check
        items:
          $ref: '#/definitions/model.AssetCoverage'
        type: array
      description:
        type: string
      evidence:
        additionalProperties: true
        description: Relevant metadata for auto-derived status
        type: object
      evidence_rules:
        description: Rules for auto-derivation
        items:
          $ref: '#/definitions/model.EvidenceRule'
        type: array
      help_text:
        description: Extended metadata fields for rich UI display
        type: string
      id:
        type: string
      info:
        $ref: '#/definitions/model.ChecklistItemInfo'
      kind:
        description: '"manual" or "auto"'
        type: string
      notes:
        description: From manual assignment
        type: string
      read_only:
        type: boolean
      recommendation:
        type: string
      required:
        type: boolean
      scope:
        description: '"global", "asset", or "business_unit"'
        type: string
      script_controlled:
        description: Can be controlled by Lua scripts
        type: boolean
      source:
        description: '"auto" or "manual"'
        type: string
      status:
        description: '"yes", "no", "na"'
        type: string
      title:
        type: string
      updated_at:
        description: From manual assignment
        type: string
      why_matters:
        type: string
    type: object
  model.EvidenceRule:
    properties:
      key:
        description: e.g., "http.title", "last_scanned_at"
        type: string
      op:
        description: '"exists", "eq", "regex", "gte_days_since"'
        type: string
      source:
        description: '"scan_metadata"'
        type: string
      value:
        description: Value for "eq", "regex", "gte_days_since"
    type: object
  model.FAQ:
    properties:
      a:
        type: string
      q:
        type: string
    type: object
...
  model.Quote:
    properties:
      source:
        type: string
      text:
        type: string
    type: object
...
  v1.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    required:
    - code
    - error
    type: object
...
  v1.SetBusinessUnitChecklistStatusRequest:
    properties:
      business_unit_id:
        description: Business unit ID
        example: bu-123
        type: string
      item_id:
        description: Checklist item template ID
        example: security-policy-001
        type: string
      notes:
        description: Optional notes
        example: Verified during security audit
        type: string
      status:
        description: 'Status: yes, no, or na'
        enum:
        - "yes"
        - "no"
        - na
        example: "yes"
        type: string
    required:
    - business_unit_id
    - item_id
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A powerful asset discovery and scanning API with Lua script support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Asset Scanner API
  version: "1.0"
paths:
...
  /checklist/asset-templates:
    get:
      consumes:
      - application/json
      description: Retrieve all asset-scoped checklist templates with their coverage
        across all assets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DerivedChecklistItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get all asset-scoped templates with coverage
      tags:
      - checklist
  /checklist/asset/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve all checklist items applicable to a specific asset with
        their current status
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DerivedChecklistItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get asset-specific checklist items
      tags:
      - checklist
  /checklist/business-unit/{businessUnitId}:
    get:
      consumes:
      - application/json
      description: Retrieve all global checklist items with their current status for
        a specific business unit
      parameters:
      - description: Business Unit ID
        in: path
        name: businessUnitId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DerivedChecklistItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get business unit checklist items
      tags:
      - checklist
  /checklist/business-unit/status:
    post:
      consumes:
      - application/json
      description: Set the status (yes/no/na) of a checklist item for a specific business
        unit
      parameters:
      - description: Business unit checklist status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.SetBusinessUnitChecklistStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Set business unit checklist item status
      tags:
      - checklist
  /checklist/coverage/summary:
    get:
      consumes:
      - application/json
      description: Get a summary of compliance coverage showing which assets are covered
        by compliance checks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get compliance coverage summary
      tags:
      - checklist
  /checklist/global:
    get:
      consumes:
      - application/json
      description: Retrieve all global checklist items with their current status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DerivedChecklistItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get global checklist items
      tags:
      - checklist
  /checklist/status:
    post:
      consumes:
      - application/json
      description: Set the status (yes/no/na) of a checklist item, either global or
        asset-specific
      parameters:
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SetStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Set checklist item status
      tags:
      - checklist
  /checklist/templates:
    get:
      consumes:
      - application/json
      description: Retrieve all available checklist item templates with covered assets
        that are not compliant (status "no")
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DerivedChecklistItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: List all checklist templates with non-compliant asset coverage
      tags:
      - checklist
  /checklist/templates/upload:
    post:
      consumes:
      - application/json
      description: Upload a JSON file containing checklist templates that will overwrite
        all existing templates
      parameters:
      - description: Templates upload request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UploadTemplatesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Upload checklist templates from JSON
      tags:
      - checklist
...
swagger: "2.0"